refactor: move theme files code to main process
