+++ src/ipc-events.ts
@@ -41,6 +41,11 @@ export enum IpcEvents {
   GET_TEMPLATE = 'GET_TEMPLATE',
   GET_TEST_TEMPLATE = 'GET_TEST_TEMPLATE',
   IS_RELEASED_MAJOR = 'IS_RELEASED_MAJOR',
+  CREATE_THEME_FILE = 'CREATE_THEME_FILE',
+  GET_AVAILABLE_THEMES = 'GET_AVAILABLE_THEMES',
+  OPEN_THEME_FOLDER = 'OPEN_THEME_FOLDER',
+  READ_THEME_FILE = 'READ_THEME_FILE',
+  GET_THEME_PATH = 'GET_THEME_PATH',
 }
 
 export const ipcMainEvents = [
@@ -60,6 +65,11 @@ export const ipcMainEvents = [
   IpcEvents.GET_TEMPLATE_VALUES,
   IpcEvents.GET_TEMPLATE,
   IpcEvents.GET_TEST_TEMPLATE,
+  IpcEvents.CREATE_THEME_FILE,
+  IpcEvents.GET_AVAILABLE_THEMES,
+  IpcEvents.OPEN_THEME_FOLDER,
+  IpcEvents.READ_THEME_FILE,
+  IpcEvents.GET_THEME_PATH,
 ];
 
 export const WEBCONTENTS_READY_FOR_IPC_SIGNAL =
