+++ src/renderer/app.tsx
@@ -3,6 +3,7 @@ import * as path from 'path';
 import { autorun, reaction, when } from 'mobx';
 
 import { EditorValues, PACKAGE_NAME, SetFiddleOptions } from '../interfaces';
+import { defaultDark, defaultLight } from '../themes-defaults';
 import { PackageJsonOptions, getPackageJson } from '../utils/get-package';
 import { USER_DATA_PATH } from './constants';
 import { ElectronTypes } from './electron-types';
@@ -12,7 +13,6 @@ import { Runner } from './runner';
 import { AppState } from './state';
 import { TaskRunner } from './task-runner';
 import { activateTheme, getTheme } from './themes';
-import { defaultDark, defaultLight } from './themes-defaults';
 import { getElectronVersions } from './versions';
 
 // Importing styles files
