+++ src/themes-defaults.ts
@@ -1,4 +1,4 @@
-import * as MonacoType from 'monaco-editor';
+import type * as MonacoType from 'monaco-editor';
 
 export interface FiddleTheme {
   name?: string;
