+++ tests/mocks/electron-fiddle.ts
@@ -10,11 +10,14 @@ export class ElectronFiddleMock {
   public arch = process.arch;
   public blockAccelerators = jest.fn();
   public confirmQuit = jest.fn();
+  public createThemeFile = jest.fn();
+  public getAvailableThemes = jest.fn();
   public getTemplate = jest.fn();
   public getTemplateValues = jest.fn();
   public getTestTemplate = jest.fn();
   public macTitlebarClicked = jest.fn();
   public monaco = new MonacoMock();
+  public openThemeFolder = jest.fn();
   public platform = process.platform;
   public pushOutputEntry = jest.fn();
   public reloadWindows = jest.fn();
@@ -27,4 +30,6 @@ export class ElectronFiddleMock {
   public showWarningDialog = jest.fn();
   public showWindow = jest.fn();
   public taskDone = jest.fn();
+  public readThemeFile = jest.fn();
+  public themePath = '~/.electron-fiddle/themes';
 }
